--- core/src/main/scala/ferjorosa/sbn/core/data/MutableDataSet.scala
@@ -1,12 +1,0 @@
-package ferjorosa.sbn.core.data
-
-/**
-  * Created by fer on 27/10/16.
-  */
-class MutableDataSet {
-
-}
-
-object MutableDataSet {
-  type MutableDataOnMemory = scala.collection.mutable.ListBuffer[DataInstance]
-}
