+++ core/src/main/scala/sbn/core/data/DataInstance.scala
