+++ core/src/main/scala/sbn/core/data/MutableDataSet.scala
@@ -1,0 +1,12 @@
+package sbn.core.data
+
+/**
+  * Created by fer on 27/10/16.
+  */
+class MutableDataSet {
+
+}
+
+object MutableDataSet {
+  type MutableDataOnMemory = scala.collection.mutable.ListBuffer[DataInstance]
+}
