+++ core/src/main/scala/sbn/core/distributions/Distribution.scala
