+++ core/src/main/scala/sbn/core/distributions/Gaussian.scala
