+++ core/src/main/scala/sbn/core/distributions/Gaussian_MultinomialParents.scala
