+++ core/src/main/scala/sbn/core/logger/Logging.scala
