+++ core/src/main/scala/sbn/core/models/BayesianNetwork.scala
