+++ core/src/main/scala/sbn/core/models/graph/DirectedGraph.scala
