+++ core/src/main/scala/sbn/core/models/graph/UndirectedGraph.scala
