+++ core/src/main/scala/sbn/core/variables/Assignment.scala
