+++ core/src/main/scala/sbn/core/variables/DistributionType.scala
