+++ core/src/test/scala/sbn/core/CustomSpec.scala
