+++ core/src/test/scala/sbn/core/distributions/GaussianSpec.scala
