+++ core/src/test/scala/sbn/core/distributions/Multinomial_MultinomialParentsSpec.scala
