+++ core/src/test/scala/sbn/core/models/BayesianNetworkSpec.scala
