+++ core/src/test/scala/sbn/core/variables/DistributionTypeSpec.scala
