+++ examples/src/main/scala/sbn.examples.core/test/test.scala
